Increase baudrate of usb port to 250k
The screen is also connected using a serial port, this means the screen
must always stay at 115200 baudrate. When increasing the default
baudrate this will modify the baudrate for all ports, including the
screen.

We use a custom GCODE command that is executed after startup to increase
the baudrate only of the usb port (port 0) and leave other ports at the
default baudrate.